Added functions

+++ CloudPlant.Service/Implementation/PlantService.cs
@@ -20,45 +20,76 @@ namespace CloudPlant.Service.Implementation
             this._plantTypeRepository = plantTypeRepository;
         }
 
-       public PlantDTO CreatePlant(PlantDTO plantDTO)
+       public PlantWithPlantTypeDTO CreatePlant(PlantCreationDTO plantCreationDTO)
         {
-            var device = _deviceRepository.GetById(plantDTO.DeviceId);
-            var plantType = _plantTypeRepository.GetById(plantDTO.PlantTypeId);
+            var device = _deviceRepository.GetById(plantCreationDTO.DeviceId);
+            if (device == null)
+            {
+                throw new Exception("Device not found");
+            }
+            var plantType = _plantTypeRepository.GetById(plantCreationDTO.PlantTypeId);
+            if (plantType == null)
+            {
+                throw new Exception("Plant type not found");
+            }
 
             var plant = new Plant
             {
-                Title = plantDTO.Title,
-                LastWatering = plantDTO.LastWatering,
-                NextWatering = plantDTO.NextWatering, //Calculate this
+                Title = plantCreationDTO.Title,
                 PlantType = plantType,
                 Device = device
             };
             _plantRepository.Insert(plant);
 
-            return (PlantDTO) plant;
+            return (PlantWithPlantTypeDTO) plant;
         }
 
         public void DeletePlant(PlantDTO plantDTO)
         {
             var plant = _plantRepository.GetById(plantDTO.Id);
+            if (plant == null)
+            {
+                throw new Exception("Plant not found");
+            }
             _plantRepository.Delete(plant);
         }
 
-        public PlantDTO EditPlant(PlantDTO plantDTO)
+        public PlantWithPlantTypeDTO EditPlant(PlantDTO plantDTO)
         {
             var plant = _plantRepository.GetById(plantDTO.Id);
+            if (plant == null)
+            {
+                throw new Exception("Plant not found");
+            }
+            var device = _deviceRepository.GetById(plantDTO.DeviceId);
+            if (device == null)
+            {
+                throw new Exception("Device not found");
+            }
+            var plantType = _plantTypeRepository.GetById(plantDTO.PlantTypeId);
+            if (plantType == null)
+            {
+                throw new Exception("Plant Type not found");
+            }
 
             plant.Title = plantDTO.Title;
             plant.NextWatering = plantDTO.NextWatering;
             plant.LastWatering = plantDTO.LastWatering;
+            plant.PlantType = plantType;
+            plant.Device = device;
 
             _plantRepository.Update(plant);
-            return (PlantDTO)plant;
+            return (PlantWithPlantTypeDTO) plant;
         }
 
-        public PlantDTO GetPlant(int id)
+        public PlantWithPlantTypeDTO GetPlant(int id)
         {
-            return (PlantDTO) _plantRepository.GetById(id);
+            Plant plant = _plantRepository.GetById(id);
+            if (plant == null)
+            {
+                throw new Exception("Plant not found");
+            }
+            return (PlantWithPlantTypeDTO) plant;
         }
     }
 }
